Renames directory for consistency
