+++ components/DataVerification/DataVerifItem.tsx
@@ -2,8 +2,8 @@ import React, { FC, useEffect, useState } from 'react';
 import { Image, ImageRequireSource, StyleSheet, Text, View } from 'react-native';
 import CustomColors from '../../styles/Colors';
 import { MasteryIcon } from '../../styles/MasteryIcon';
-import { ChainStep } from '../../types/CHAIN/ChainStep';
-import { ChainStepPromptLevel, StepAttempt } from '../../types/CHAIN/StepAttempt';
+import { ChainStep } from '../../types/chain/ChainStep';
+import { ChainStepPromptLevel, StepAttempt } from '../../types/chain/StepAttempt';
 import { DataVerificationControlCallback } from '../../types/DataVerificationControlCallback';
 import BehavAccordion from './BehavAccordion';
 import BehavDataVerifSwitch from './BehavDataVerifSwitch';
