+++ screens/ProbeScreen.tsx
@@ -8,10 +8,10 @@ import { DataVerificationListItem } from '../components/Probe/';
 import { ImageAssets } from '../data/images';
 import { ApiService } from '../services/ApiService';
 import CustomColors from '../styles/Colors';
-import { ChainSession, ChainSessionType } from '../types/CHAIN/ChainSession';
-import { ChainStep } from '../types/CHAIN/ChainStep';
-import { ChainData } from '../types/CHAIN/SkillstarChain';
-import { ChainStepStatus, StepAttempt } from '../types/CHAIN/StepAttempt';
+import { ChainSession, ChainSessionType } from '../types/chain/ChainSession';
+import { ChainStep } from '../types/chain/ChainStep';
+import { ChainData } from '../types/chain/SkillstarChain';
+import { ChainStepStatus, StepAttempt } from '../types/chain/StepAttempt';
 
 const ProbeScreen = (): JSX.Element => {
   const navigation = useNavigation();
