+++ types/Context.ts
@@ -1,7 +1,7 @@
-import { ChainSession } from './CHAIN/ChainSession';
-import { ChainStep } from './CHAIN/ChainStep';
-import { MasteryInfo } from './CHAIN/MasteryLevel';
-import { ChainData, SkillstarChain } from './CHAIN/SkillstarChain';
+import { ChainSession } from './chain/ChainSession';
+import { ChainStep } from './chain/ChainStep';
+import { MasteryInfo } from './chain/MasteryLevel';
+import { ChainData, SkillstarChain } from './chain/SkillstarChain';
 import { Participant, User } from './User';
 
 export type ContextStateValue =
